requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
fake-useragent==1.4.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0 